feat(pipeline/stages): 新增所有核心流水线阶段处理器
添加交易引擎全套流水线处理模块，包括市场数据处理、订单管理、执行逻辑、策略调用、账户统计与订单清理功能，同时提供通用工具函数与完整单元测试覆盖。
